+++ eslint-config/rules/plugins/import.js
@@ -79,8 +79,8 @@ module.exports = {
         ],
         'import/newline-after-import': 'error', // enforce a newline after import statements
         'import/no-default-export': 'off', // allow default exports
-        'import/prefer-default-export': 'error', // prefer a default export if module exports a single name
         'import/group-exports': 'off', // allow export statements anywhere
+        'import/prefer-default-export': 'warn', // prefer a default export if module exports a single name
         'import/max-dependencies': 'off', // limit the maximum number of dependencies a module can have
         'import/no-unassigned-import': 'off', // forbid unassigned imports
         'import/no-named-default': 'error', // forbid named default exports