+++ eslint-config/test/index.js
@@ -18,8 +18,8 @@ test.beforeEach((t) => {
 
 test('ESLint config fails for improperly formatted .js file', (t) => {
     const report = t.context.cli.executeOnFiles([ 'data/bad.js' ]);
-    const numErrors = 19;
-    const numWarnings = 1;
+    const numErrors = 18;
+    const numWarnings = 2;
 
     t.is(report.errorCount, numErrors);
     t.is(report.warningCount, numWarnings);